backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
shap
apscheduler
pytest
requests
PyJWT
